Making login system work with routing
The login system currently works by user intering the username and password. Then, it get the user with that username. If such a username exists then it gets gets the password and it hashes it using md5, if the password is correct then the website is navigated to the details page"

making navigation bar only appear when logged in

Making the login System work

The login system now works if the user enters the right username and password. The user can't go to routes unless they are signed in. If the user is not signed in then they are directed to the login page. I also made a module for the shared features and the navigation bar components needs to be shared between different modules. So it exports the navigation bar to other modules.

the login works